+++ docs/history.rst
@@ -55,8 +55,8 @@ Backwards Compatibility Notes
 * The ``pyembed::OxidizedPythonInterpreterConfig`` Rust struct had its
   ``packed_resources`` field changed from ``Option<&'a [u8]>`` to
   ``Vec<&'a [u8]>`` so multiple resource inputs can be specified.
-* The ``PythonDistribution`` Starlark type no longer has a
-  ``package_resources()`` method. Use
+* The ``PythonDistribution`` Starlark type no longer has
+  ``package_resources()`` and ``source_modules()`` methods. Use
   ``PythonDistribution.python_resources()`` instead.
 
 New Features